+++ tests/tt_eager/python_api_testing/unit_testing/misc/test_flash_multi_latent_attention_prefill.py
@@ -23,8 +23,6 @@ from models.tt_transformers.tt.common import (
     PagedAttentionConfig,
 )
 
-ENABLE_SDPA_STRESS_TESTS = os.getenv("ENABLE_SDPA_STRESS_TESTS", "0") == "1"
-
 
 def nearest_n(x, n):
     return ((x + n - 1) // n) * n
@@ -217,7 +215,6 @@ def run_flash_mla_prefill_impl(
     assert out_pass, f"Output mismatch: PCC {out_pcc} < 0.99"
 
 
-@pytest.mark.skipif(ENABLE_SDPA_STRESS_TESTS, reason="Only running stress tests")
 @pytest.mark.parametrize(
     "batch, seq_len, nh, nkv, kv_lora_rank, d_rope",
     # batch, seq_len, num heads q, num heads kv, kv lora rank, dim rope
@@ -251,102 +248,6 @@ def run_flash_mla_prefill_impl(
     ],
 )
 def test_flash_mla_prefill(
-    device,
-    batch,
-    seq_len,
-    nh,
-    nkv,
-    kv_lora_rank,
-    d_rope,
-    q_dtype,
-    dtype,
-    use_paged_attention,
-    block_size,
-    function_level_defaults,
-    reset_seeds,
-):
-    run_flash_mla_prefill_impl(
-        device,
-        batch,
-        seq_len,
-        nh,
-        nkv,
-        kv_lora_rank,
-        d_rope,
-        q_dtype,
-        dtype,
-        use_paged_attention,
-        block_size,
-    )
-
-
-@pytest.mark.skipif(not ENABLE_SDPA_STRESS_TESTS, reason="Stress tests are disabled by default.")
-@pytest.mark.parametrize(
-    "batch",
-    [
-        1,  # Single batch
-        # 2,  # Multiple batches # Removing to reduce CI load
-        8,  # Even larger batch size
-    ],
-)
-@pytest.mark.parametrize(
-    "seq_len",
-    [
-        1 * 1024,  # Long sequence length
-    ],
-)
-@pytest.mark.parametrize(
-    "nh",
-    [
-        16,
-        32,
-        128,
-    ],
-)
-@pytest.mark.parametrize(
-    "nkv",
-    [
-        1,
-        # 8, # Removing to reduce CI load
-        16,
-    ],
-)
-@pytest.mark.parametrize(
-    "kv_lora_rank",
-    [
-        64,
-        512,
-    ],
-)
-@pytest.mark.parametrize(
-    "d_rope",
-    [
-        0,
-        32,
-        128,
-    ],
-)
-@pytest.mark.parametrize(
-    "q_dtype, dtype",
-    [
-        (ttnn.bfloat16, ttnn.bfloat8_b),
-    ],
-)
-@pytest.mark.parametrize(
-    "use_paged_attention",
-    [
-        False,
-        True,
-    ],
-)
-@pytest.mark.parametrize(
-    "block_size",
-    [
-        32,
-        128,
-    ],
-)
-def test_flash_mla_prefill_stress(
     device,
     batch,
     seq_len,
